// src/lib/types/nation.ts
export interface GovtPercentages {
    admin: number;
    defense: number;
    education: number;
    environment: number;
    healthcare: number;
    commerce: number;
    foreignaid: number;
    lawenforcement: number;
    publictransport: number;
    socialequity: number;
    spirituality: number;
    welfare: number;
}
export interface DeathPercentages {
    age?: number;
    hdisease?: number;
    murder?: number;
    lost?: number;
    god?: number;
    cancer?: number;
    exposure?: number;
    suicideIPC?: number;
    accident?: number;
    animal?: number;
    disappearance?: number;
    work?: number;
    ritual?: number;
    war?: number;
    punishment?: number;
    bungee?: number;
    scurvy?: number;
    nuclear?: number;
    euthanasia?: number;
    sunburn?: number;
    vat?: number;
    malnourishment?: number;
    space?: number;
}

export interface Sectors {
    blackmarket?: number;
    government?: number;
    industry?: number;
    public?: number;
}